pyproject.toml
[tool.poetry]
name = "pythonprojects"
version = "0.1.0"
description = ""
authors = ["Adam <adamrek.st9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
rich = "^13.9.4"
tqdm = "^4.67.0"
ascii_graph = "^1.5.1"
pillow = "^11.0.0"
numpy = "^2.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
